Styling updates
Moved all styled components to a single utility component.
