+++ src/components/footer.js
@@ -2,41 +2,20 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Img from 'gatsby-image';
 import { graphql, useStaticQuery } from 'gatsby';
-import styled from '@emotion/styled';
 
 import { ReactComponent as GitHub } from '../images/github.svg';
 import { ReactComponent as Twitter } from '../images/twitter.svg';
 import { ReactComponent as RSS } from '../images/rss.svg';
 import {
-  absolute,
-  aSelfCenter,
-  backgroundAccentLight,
-  bottom0,
-  content,
-  contentBetween,
-  contentCenter,
-  flex,
-  flexColumn,
-  f4,
-  f5,
-  grid,
-  itemsCenter,
-  itemsFlexStart,
-  itemsStart,
-  jSelfEnd,
-  left0,
-  ma0,
-  mb1,
-  mHorizontal,
-  normal,
-  noUnderline,
-  overflowHidden,
-  pa2,
-  relative,
-  right0,
-  w100,
-  zUnder,
-} from './util/styleUtils';
+  DD,
+  DL,
+  DT,
+  FootBody,
+  FooterCopy,
+  FooterMain,
+  FootHead,
+  SocialLink,
+} from './util/styleEl';
 
 const size = {
   height: '6.5em',
@@ -50,123 +29,6 @@ const imgStyles = {
   ...size,
   borderRadius: '50%',
 };
-
-const FooterMain = styled.footer`
-  ${contentCenter};
-  ${grid};
-  ${itemsStart};
-  ${mHorizontal};
-  ${pa2};
-  ${relative};
-  ${w100};
-
-  background-color: var(--bg-dark);
-  border-color: var(--primary);
-  border-style: solid;
-  border-width: 1rem 0 0;
-  grid-column: 1/4;
-  grid-gap: 1em;
-  grid-row: 3/4;
-  grid-template-columns: 6.5em minmax(10.5em, 32em);
-  grid-template-rows: min-content 2.5em;
-
-  &::before {
-    ${absolute};
-    ${backgroundAccentLight};
-    ${bottom0};
-    ${content};
-    ${left0};
-    ${right0};
-    ${w100};
-    ${zUnder};
-
-    top: -2rem;
-    min-height: 1rem;
-  }
-`;
-
-const FooterCopy = styled.section`
-  ${flex};
-  ${flexColumn};
-  ${itemsFlexStart};
-  ${mb1};
-  ${normal};
-  ${w100};
-
-  max-width: 32rem;
-
-  &,
-  &:hover,
-  &:visited {
-    ${f4};
-
-    color: hsl(64, 32%, 96%);
-  }
-`;
-
-const FootHead = styled.h3`
-  ${f4};
-  ${ma0};
-  ${mb1};
-`;
-
-const FootBody = styled.p`
-  ${f5};
-
-  max-width: 32rem;
-`;
-
-const DL = styled.dl`
-  ${aSelfCenter};
-  ${contentBetween};
-  ${flex};
-  ${itemsCenter};
-  ${jSelfEnd};
-  ${ma0};
-  ${w100};
-
-  grid-column: 2/3;
-  grid-row: 2/3;
-  max-width: 8rem;
-`;
-
-const DT = styled.dt`
-  ${absolute};
-  ${overflowHidden};
-
-  clip: rect(0.1rem, 0.1rem, 0.1rem, 0.1rem);
-  height: 0.1rem;
-  width: 0.1rem;
-`;
-
-const DD = styled.dd`
-  ${ma0};
-
-  height: 2.5em;
-  width: 2.5em;
-`;
-
-const SocialLink = styled.a`
-  ${contentCenter};
-  ${flex};
-  ${itemsCenter};
-  ${noUnderline};
-  ${relative};
-
-  height: 2.5em;
-  width: 2.5em;
-
-  svg {
-    fill: #fff;
-    transition: fill var(--duration) ease-in-out;
-  }
-
-  &:hover {
-    svg {
-      fill: var(--primary-light);
-    }
-  }
-`;
 
 const Footer = ({ description, social }) => {
   const {
